refactoring the core benchmarks and add pow benchmarks
